autotest: Fix problem with logging_UserCrash tgz file being masked.
We use a tgz file to hide a binary file that we build and don't want
portage to strip.  With a change yesterday, autotest-test ebuild
started removing all tgz files.  Rename it to avoid detection.  We
could improve this test by running dump_syms on the host and
populating a breakpad directory there to avoid needing to sneak
symbols onto the target, but that's a larger change.

BUG=7009
TEST=Run logging_UserCrash

Change-Id: Ib6a131c3656d3ae1a8f96cac49dfe29e3ab71b33

Review URL: http://codereview.chromium.org/3450025