add per-step failure support for WiFi tests
Each step in a test can now be prefixed by "!" to indicate the step will
fail.  This is a convenient mechanism for testing connection requests
that should fail (e.g. due to invalid parameters).  We should check the
error thrown is what we expect but right now that's painful.

BUG=none
TEST=035CheckWEPKeySyntax (coming in another CL)

Review URL: http://codereview.chromium.org/3405016

Change-Id: Ic62b63cde8ec5e8f773d8093bbe2997e820cfda4