Suspend function for site_wifitest
Suspend for a specified number of seconds.  This makes use a if a newly created python script for using the RTC to trigger a suspend, then running a command.  There is both a foreground and background variant.

BUG=none
TEST=Created a test in the Roaming Suite (not yet approved).
Manually ran command on x201, l13 and Andretti.

Review URL: http://codereview.chromium.org/3471015