Early days of WiFi Roaming Test Suite
This is a first pass at a WiFi test suite aimed at roaming and
general network stability over changing conditions.  Currently
the two tests we do are crude and involve some fairly crude
manipulation of the servers behind the back of the supplicating
DUT.

First, in 000ChannelHop we move the AP between channels, and
watch to see if the DUT moves over and reconnects.

Second, in 001SwitchBack, we connect to two distinct SSIDs
(so that both are in the profile), then we reconfigure the
AP back to the old SSID without informing the DUT.

In order to monitor the DUT, changes are made to site_wlan_connect
to optionally passively watch for a connection without actually
commanding the DUT to do anything.  Changes were also made to
site_wifitest.py to pass these arguments in from a "connect" step.

BUG=none
TEST=run_remote_tests

Review URL: http://codereview.chromium.org/3439004