Improve accuracy of wait_state after suspend
The output from the current Suspend tests are very inaccurate
since the event wait script runs as a separate SSH instance that occurs
after the system wakes up (i.e, after USB-Ethernet resumes from
suspend and the TCP state comes back and closes the old SSH connection
that did the suspend.  This leads to wildly inaccurate numbers.
This change causes the suspend and wait to occur right after each other,
and as a result the numbers are VASTLY more consistent.

This change breaks open the wait_service function in site_wifitest
into two pieces -- the part that generates command line arguments
to run on the client, and the other half to read back the result.
There's a trivial third function that replicates the old function.

Two new functions are created that run the wait_service function
after a suspend-resume cycle.  The former function starts the
suspend-resume and passes the command line for wait_service so
that wait_service is executed immediately after suspend, so we don't
have the ethernet re-acquisition time randomizing the test result.
The second function collects the information from the helper thread
that ran the suspend-resusme-wait_service in the background and passes
it to the wait_service output parser.

Along the way I had to modify the suspend script a little bit since
in some code paths the suspend-to-ram code will output things to
stdout, and stdout is reserved for data output from the after-proc.

BUG=none
TEST=Repeated runs

Review URL: http://codereview.chromium.org/3521004