[autotest] Remove created cryptohome as a part of test cleanup.
With the new parallel login, if we have a cryptohome on-disk that we
can mount, login succeeds -- even if online login is going to fail.
Thus, in order to force the BadAuthentication tests to continue to
test the same code pathway in the new code that they're testing in the
old code, we need to make sure that we have no cryptohome on disk for
the test user.

BUG=chromium-os:4929
TEST=BVT

Change-Id: I40df641c7b2794fb0567f03703cf9400a5f8109d

Review URL: http://codereview.chromium.org/3571016