Optional 3rd step argurment to search for error messages
When a test that has a "!name" step occurs, we previously considered
the test successful.  We now add another optional critera which is
a regular expression to search for in the error output.  This desired
output is specified as a third parameter in the step list (which
previosly only contained the "!step" name and dict containing parameters
for the test) containing the regular expression to search for in the
error string.

The error string can be specified by the step -- i.e,
"self.error_string" set inside the test before raising the exception, or
more commonly, it can be retrieved from the CmdError exception's stdout.
For example host.run(cmd) will return an exception if the ssh command
fails and the stderr of the function call will be available inside the
exception.

I also found and removed some semicolons.  Let me know if
you'd like that broken off into another CL.

BUG=none
TEST=Run on working tests (SecMat.020CheckRSN), run on current "!connect"
tests (SecMat.035CheckWEPKeySyntax), and run on a new "!connect" test
that specifies a search error string (SecMat.021CheckWPA_BadPassphraseLength)
new tests which have the third argument

Review URL: http://codereview.chromium.org/3668001