Check the existence of the cardreader using log messages.
The cardreader is always power off until a card inserted. The previous
approach using lspci needs to insert a card to activate it while the
detection. It is not so effective because the component validation has
no operator-attended.

So the new approach is to check the log messages. If a carder is
inserted before the test, that inform can be record there. The above
assumption is feasible because there is an interactive test for the
cardreader.

BUG=chrome-os-partner:877
TEST=tested on non-Mario device

Change-Id: I58b817c739e62118b6427f6de79ca3459dee64ac

Review URL: http://codereview.chromium.org/3584013