Make our http(s) server multithreaded
A user-agent like chrome may make simultaneous connections to an HTTPS server, and then send data only on the one that the server believes to be the second.  If the server is blocking waiting for data on the first connection, it gets hung.  Our test https server was doing just that.  So, make it multithreaded.

Change-Id: I537fb633e236b826f85752b206397831b9124ea8

Revert "[autotest] Temporarily remove login tests from suite_Smoke"

This reverts commit 353c17eba46b80c19b15e8a9c0a8635619f3262f.

BUG=8383
TEST=suite_Smoke (with login tests added back in) works again!

Review URL: http://codereview.chromium.org/4205006