Enhance BT tests to check inputs from different devices.
This change adds to the standard page generator (in
site_ui.py) the ability to generate checkboxes and text
fields as requested by the user.

The BluetoothSemiAuto rtest is being enhanced to allow the
operator to indicate that all input devices (Bluetooth and
build-in mouse and keyboard) can coexist.

If a checkbox is left unchecked or a text field is left
empty, the test fails.

Change-Id: I7733758ec1f1f536fe9fb2dad198a06ddcea64ac

BUG=chromium-os:3695, chromium-os:3696
TEST=see below:

1. Ran
./run_remote_tests.sh  --remote=172.22.75.138 BluetoothSemiAuto

Observed that it is still possible to connect with a BT
mouse (did not try connecting with a BT keyboard as I don't
have one). If both checkboxes are checked and bot text input
fields have some characters in them, the auto test passes.

If one of these conditions is not met - the test fails
reporting the problem as appropriate:

vvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvv
TestFail: Bluetooth input errors:
no input in "BT Keyboard" field
no input in "Built In Keyboard" field
^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
2. As a regression test ran

 ./run_remote_tests.sh  --remote=172.22.75.138 TearTest
 ./run_remote_tests.sh  --remote=172.22.75.138 SanAngeles

Review URL: http://codereview.chromium.org/5194001