Improve BT keyboard pairing test.
It was discovered that the 'Microsoft Mobile Keyboard 6000'
fails to pair when running the BluetoothSemiAuto test. The
Logitec MX5500 keyboard on the other hand was connecting as
expected. The difference is that the Logitec keyboard has a
display where it indicates when the user is expected to
enter the pin code, whereas the Microsoft keyboard is
'blind'.

It turned out that the problem with the Microsoft keyboard
was caused by wrong synchronization: the user would not
enter the PIN code, or would enter it too soon before the
DUT is ready.

This CL adds a page which splashes several seconds before
the DUT is ready to pair. The operator is expected to enter
the PIN 5 seconds after this splash page  closes.

Change-Id: I25b5b6d509dbdd81f39ca27485ddebc375772650

BUG=chromium-os:3696
TEST=see below

Bring a BT mouse and a BT keyboard into the 'waiting to
pair' state and from src/scripts directory on the host run

 ./run_remote_tests.sh  --remote=172.22.75.33 BluetoothSemiAuto

The test proceeds as before, but when user clicks on a
keyboard address to pair it,  a new screen shows up briefly
(for 5 seconds), requesting user to enter '0000' (that's
four zeros) on the keyboard to be paired 5 seconds after the
page closes.

Once the user waits 5 seconds and enters 0000<cr> on the
keyboard it pairs and its input is available when the next
screen is displayed.

The mouse pairs as before, no intermediate screen is
displayed.

Once both mouse and keyboard are paired and the appropriate
input actions are completed, the test reports success.

INFO   : Test results:
-----------------------------------------------------------
hardware_BluetoothSemiAuto                            PASS
hardware_BluetoothSemiAuto/hardware_BluetoothSemiAuto PASS
-----------------------------------------------------------
Total PASS: 2/2 (100%)

Review URL: http://codereview.chromium.org/5306006