Fix "bug: sequence status mixes old and new state".
Actually pay attention to the iteration count when processing
the status for automated sequences.  Before this was the
intention, but that code never executed.

BUG=chrome-os-partner:1623
TEST=manual

To test:

Define an automated sequence, say A B C, such that A and B pass
but C always fails.  Before if you run the seq a second time, it
will not run B because it sees the prev C failure.  Now it should
always run all test on each repeated call.

Change-Id: I681095c9ca050292a38087e622230593e504dcbe

Review URL: http://codereview.chromium.org/5353002