Add an extra reboot to perfalerts test to work around first boot anomaly
If the install procedure involves stateful_update (which it does in
the test lab prior to running perfalerts test), the first boot of
the test will spend an extra second or more mounting the stateful
partition.  This change works around the problem by booting the
target once before starting the test.  This has the effect of
throwing away the single long boot time.

Change-Id: Iabf0618739ca9b7276df8578b5c6be750ed89c29

BUG=chromium-os:10392
TEST=run_remote_tests.sh suites/control.perfalerts

Review URL: http://codereview.chromium.org/5926002