Convert seconds_kernel_to_login to use the "boot-complete" event.
Recent changes in the session manager, window manager, and
boot-complete upstart job changed the timing and meaning of
the "login-prompt-ready" bootstat event, and added a "boot-complete"
event.  This change converts the seconds_kernel_to_login perf
keyval to use the "boot-complete" event, which is the best
available proxy for the event that marks when the user will see
the system as booted and ready.

The change also adds a rdbytes_kernel_to_x_started for symmetry
with seconds_kernel_to_x_started, and deletes the now obsolete
sectors_read_kernel_to_login perf keyval.

BUG=chromium-os:10686
TEST=platform_BootPerfServer (via bootperf-bin/bootperf)

Review URL: http://codereview.chromium.org/6062008

Change-Id: I9f17993fef4d44220e36a45d033b45ca7e0f5c59