autotest: introduce flimflam_test_path.py, make use of it in all the places
Flimflam needs path magic to be imported, the new file unifies the way
in which the path is determined across all places, and eliminates some
copy/paste.

Take two of the CL, trying to be less radical and ground(bot)breaking.
Instead of trying to wrap around the whole package, I'm wrapping just
around exporting the path, and flimflam import is done in the same place
as before, to keep namespaces intact.

BUG=11138
TEST=ran BVT's

Change-Id: I6c4abfd7503a1afb0ac81a94024546717d833d3e

Review URL: http://codereview.chromium.org/6269017