Modify GPS test to work with gpsd.
This CL modifies the test to work with gpsd and loosens the
requirements for the GPS output: we don't expect it to be
fixed, we just observe that the NMEA stream is coming from
the device.

This is not a thorough test, but it is better than nothing,
a much more sophisticated setup (including a box simulating
satellite signals) would allow to make the test more
comprehensive.

This test is not yet being included in the list of factory
tests, as a tweak to /etc/init/gpsd.conf is required for
proper `initctl' operation.

Change-Id: I263e98afc3da860f95516dbfbe30aef28308cd3f

BUG=chromium-os:467
TEST=see below:

. build system image
. modify it for test
. install on the device under test
. on the desktop is src/scripts run

./run_remote_tests.sh --board=x86-mario --remote=<taregt ip> hardware_GPS/control

. observe the results:

INFO   : Test results:
-------------------------------
hardware_GPS              PASS
hardware_GPS/hardware_GPS PASS
-------------------------------
Total PASS: 2/2 (100%)
Elapsed time: 0m14s

. try the above with gpsd running and stopped on the
  target, make sure the test succeeds in both cases and the
  gpsd status is not changed by the test

Review URL: http://codereview.chromium.org/6286034