Add Gmail Latency test for Chrome OS
This is a small first step to measuring Gmail peformance on Chrome OS. It
is based on the UrlFetch test and uses a known URL to run the test. At the
moment it is not suitable for running as part of the test suite, nor for
external use (outside Google) as it needs a special server.

This is a stop-gap test to get the Gmail team going. I expect to replace
it with a PyAuto variant by the end of Feb11.

It requires an enhanced url_fetch_test binary, with JavaScript waiting
capability, which is included in this test.

BUG=chromiumos-10869
TEST=run test on x86-mario, qemu. Does not work on ARM.

Change-Id: Ifd0fbf8cfb463405dad8dfa9a7800e0541d9a27a

Review URL: http://codereview.chromium.org/6133009