Supress an annoying message from the autoserv console output.
DEBUG:root:unable to import site symbol '_SiteRun', using non-site implementation

Basically, if there is an site_autotest module under server directory, it will try to import a class named _SiteRun() out of it. If it is not there, this message will be logged out to the console, harmless, but quite annoying and misleading new users.

So I just add a dummy class there.

Change-Id: I946b2a7022a41c9ed2d00a5ef58bd4f03834e921

BUG=
TEST=run_remote_test with sleeptest.

Review URL: http://codereview.chromium.org/6524027