autotest: fix WiFi frequency handling
The site_wlan_connect script used to establish a WiFi connection did
not handle a station roaming behind it's back.  Specifically the
frequency reported for an association might not be collected causing
the script to raise an error.  Guarantee the reported frequency is
always a number and collect network state on any service state change;
this handles roaming (direct transition to "ready" state) and guards
against reporting a frequency as "None"

BUG=none
TEST=run network_WiFiPerf/060Perf24HT20

Review URL: http://codereview.chromium.org/6576010

Change-Id: I47a35ec1540bd74df60b3b035d3cc3d6a1d8b62e