power_CPUFreq option to set one frequency for all cpus
power_CPUFreq autotest was failing on ARM because it was attempting to
set the CPU core frequencies independently, and ARM apparently does not
support that.  This patch gives the test a second chance: it sets the
frequency for all cores at the same time.

BUG=chromium-os:12455
TEST=Run power_CPUFreq on both ARM and x86, should pass.

Change-Id: Ie9d281d70115fa9d00a114f9f80c88d54a56e9da
Signed-off-by: Simon Que <sque@chromium.org>

Review URL: http://codereview.chromium.org/6575021