Fix kernel error paths test.
This CL modifies the platform_KernelErrorPaths autotest to
properly utilize the capabilities provided by the
/proc/breakme interface on the target (the source file of
this interface can be found in
http://git.chromium.org/gitweb/?p=kernel.git;a=blob;f=fs/proc/breakme.c.

For nmi watchdog to operate properly, the kernel command
line needs to be modified (see http://codereview.chromium.org/6597001)

Few of the strings handled by this 'breakme' interface
('deadlock', 'softlockup', and 'irqlockup') do not cause
target reset, even though some of them do lock it up.

This will have to be investigated and fixed on the kernel
side, then the strings could be added to this test.

Change-Id: If62b0b0baef9176105ad59ff1ef1dda75a5e17c6

BUG=chromium-os:12463, chromium-os:12464
TEST=see below:

== 1 ==

run in chroot ~/trunk/src/scripts:

 ./run_remote_tests.sh --board=x86-mario --remote=<target_ip>  platform_KernelErrorPaths

observe the target reset a few times, watch for

INFO   : Test results:
---------------------------------------------------------
platform_KernelErrorPaths                           PASS
platform_KernelErrorPaths/platform_KernelErrorPaths PASS
---------------------------------------------------------
Total PASS: 2/2 (100%)

in the end of the output.

The test takes 157 seconds to complete.

== 2 ==

. repeat the above test twice, once with no file
'/home/chronos/Consent To Send Stats' present on the target
and again with the file present and containing some
multiline text.

After each test run the state and contents of the file are
maintained.

==

It looks like there is a problem with timeout handling by
autotest (in cases when the target does not restart, or is
not available at the beginning of the test) - this needs to
be investigated separately.

Another autotest enhancement could be shortening the time it
takes to reestablish ssh connection after target restarts.

Review URL: http://codereview.chromium.org/6596002