[autotest] Add autotests for the ownership API calls
most of the code here is to create a key pair we can use to create the
signatures used in the ownership API.  We leverage command-line tools, because
we also need to inject keys into the user's real NSS database.

At that point it's easier to just use openssl command line tools to generate
the necessary signatures.

This first pass just tests Whitelist/Unwhitelist.

BUG=6087
TEST=this is a test.  run it :-)

Change-Id: I1440558543a4f52b93cd565db3d433e61605075d

Review URL: http://codereview.chromium.org/6601001