Merge remote branch 'autotest-upstream/master' into merge-upstream
rev 5270 and 5271.

BUG=None
TEST=Run BVT test.

Change-Id: I46ee56cea4861d9dd35b98f0a21536122714518c

[PATCH] After_iteration hooks are not executed if test fails.

Base_test's _call_run_once() does not execute the after_iteration hooks
if the test fails.  It is important to execute the after_iteration hooks
in the event of a test failure since these hooks may be used to gather
important test logs to help diagnose the failure.  This patch
modifies _call_run_once() to execute after_iteration hooks regardless of
the outcome of the test.

A new unit test was added to verify the behavior of _call_run_once().  The
patch was tested using unittest_suite.py to ensure that it did not break
other functionalities.  It was also tested with an actual after_iteration
hook that collected test logs and the test was forced to fail.

Risk: High
Visibility: The order of when the after_iteration hooks has been changed
and may affect existing tests.
Bugs: None

Signed-off-by: Thieu Le <thieule@chromium.org>

git-svn-id: svn://test.kernel.org/autotest/trunk@5271 592f7852-d20e-0410-864c-8624ca9c26a4

KVM test: Make unattended setup copy finish.exe to floppies

Fixing yet another bug from the conversion of unattended
install to infrastructure.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>

Review URL: http://codereview.chromium.org/6598041