Introduce Trackpad Autotest framework -- hardware_Trackpad module
This module focuses on the high level procedures, such as
- read functionality list to test
- open device file for reading trackpad packets and play them back
- Initiate Xcapture module to capture X events
- Initiate Xcheck module for checking if X events meet the criteria for a
  specific functionality
- Iterate though every functionality, and for each functionality, iterate
  through every gesture file
- Calculate statistics of test results

This module is not testable yet. There are two modules missing at this time
for easier review, i.e., Xcheck.py, and Xcapture.py. Xcheck is used to check
X events and will grow larger later when all functionalities are supported
in the future. Hence, it is better to separate it as a module by itself.
Those two modules above will be uploaded later.

BUG=chromium-os:14060
TEST=None

Change-Id: I1bd58ce3a06829f29096f56feac79944c186c0ee
Reviewed-on: http://gerrit.chromium.org/gerrit/2286
Reviewed-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>