Introduce trackpad autotest framework - recording gestures
In this CL, only record program is uploaded for review. In another CL, the
autotest program, i.e., hardware_Trackpad.py and its control file will be
uploaded.

Only 3 functionalities are recorded in the list. We will expand the list
gradually.

BUG=chromium-os:14060
TEST=Follow steps below
On a host
(chroot) $ ./start_devserver

On a chromeos machine
$ gmerge utouch-evemu
$ sudo python trackpad_record

Change-Id: Icb51599d8dc2242138384986c1745a34ae24f8d3
Reviewed-on: http://gerrit.chromium.org/gerrit/329
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>