Update PlaybackRecordSemiAuto test to use Alsa
The test_tones executable was updated to output audio using Alsa instead of PulseAudio.  The Python script to run the test was also updated to gather information from Alsa and use Alsa commands to play and record audio.

Alsa does not have the notion of a 'port' (e.g. Headphone, Speaker, etc.) but instead  has volume controls for such things as Headphone, Speaker, Capture, etc.  Tests can be easily added to test various configurations in the TESTS_ declaration.

The script needs some cleanup, but it is fully functional now so I wanted to get some feedback.

When this is checked in, the pulseaudio dependency can be removed from the autotest ebuild with CL 6628018.

Change-Id: If736d849e0f53efe3e5ec49ba346d5b386dd9419

BUG=none
TEST=Test will run without PulseAudio installed

Review URL: http://codereview.chromium.org/6594003