[autotest] Keep up with the varying needs of chrome testing binaries
Since last we met, browser_tests no longer deals well with being run
as root.  So, run it as chronos.  Also, the little test server spun up
as a part of many tests wants to write a log file in cwd, so the cwd
has to be writable by chronos.

BUG=8544
TEST=run desktopui_BrowserTest.* and see that the tests all run, though they don't all pass

Change-Id: I130ff87e43138c41847d417478ab9945cf0a8876

Review URL: http://codereview.chromium.org/6676004