Create new test suite named "network_WiFiManagerTests"
This new test suite is meant for tests that are aimed at the connection
manager, as opposed to the driver / networking hardware / firmware.
Start by moving tests out of SecMat and MatFunc which were of this type
so vendors can concentrate on tests that directly affect them.

BUG=chromium-os:13033
TEST=Run new suite

Change-Id: Ic1d559dad02b95c80c05727bf089c9a4456f0e22

Review URL: http://codereview.chromium.org/6679005