kernel_Lmbench: initial commit of test
Copied the test over from client/tests. The upstream test relies on make
and perl. I wanted to avoid that. I also wanted to run each benchmark
individually so that we can tune the runs to be as deterministic as
possible (using taskset, nice, etc). I also wanted to be able to
break up the test so that we can run individual tests.

BUG=12233
TEST=Ran the test.

Change-Id: I42ecccd87b43b509ba02ea0925b52935976ea14b

Review URL: http://codereview.chromium.org/6598082