Fix realtimecomm_GTalk[Audio]Playground tests.
Updated the GTalkPlayground.tar.gz file to have the correct relative
paths (the last committed version was created from a different root
directory)

Modified tests to use "talk.google.com" as the URL, and rely on the DNS
redirection to point it to the localhost server. Also, the server now
runs on port 80 (so the hostname matching logic in GoogleTalkPlugin
whitelists it), which requires running with empty credentials (guest),
which disables the fake auth service that usually runs on port 80.

Updated the logic for getting pids to handle process names longer than
15 characters. It turns out that /proc/{pid}/stat only contains the
first 15 characters of each process name, and GoogleTalkPlugin is 16
characters long.

Made the frame count logic in realtimecomm_GTalkPlayground tests check
the number of expected frames against a function of the specified test
time, instead of a hardcoded number. In the future, if the test time is
shortened, the test should continue to pass without further
modification.

Removed the option to allow access to *.corp.google.com, since the test
is running a local server anyway and can use any valid URL to match
against it (with the DNS redirection).

Updated the log location to /tmp/gtalkplugin.log. Though it appears that
this value was overwritten in the 'options' files to '/tmp/tmp.log', it
had no apparent effect when running the tests.

Also fixed typo in cros_ui_test.py (unrelated).

Change-Id: I458e54486db03471f1835baae2b0a3f8c2f98a9d

BUG=chromium-os:3001, chromium-os:12117
TEST=Ran tests on Cr48.

Review URL: http://codereview.chromium.org/6713133