Add new site parser based off on generate_test_report.
This site parser is used to generate a JSON report of test failures, crashes,
and the associated logs for later consumption by an Email generator.

The parser uses the test report generator which comes bundled with the Chrome
OS source tree in order to maintain consistency. As well as not having to keep
track of any secondary failure white lists.

The path to the Chrome OS source tree is defined in global_config under the
CROS section as 'source_tree'.

Existing parse behavior is kept completely intact. If the site parser is not
configured it will print a debug message and exit after default parser is
called.

Relies on changes to generate_test_report here:

http://codereview.chromium.org/6731024/

Change-Id: Ib4696b88a22ca67b0822a72366d57709a5863ae5

BUG=chromium-os:12100, chromium-os:13586, chromium-os:13496
TEST=Ran parser on many sets of results.

Review URL: http://codereview.chromium.org/6758040