Modify CheckIBSS to use local DHCP server on AP
The previous version of CheckIBSS used an unsupported configuration:
using the software bridge with a wireless interface in IBSS mode.
Newer versions of OpenWRT expressly forbid this configuration.
Instead, provide a facility for running a local DHCP server on the
access point for purposes of testing the IBSS mode.

BUG=chromium-os:13871
TEST=Run test on both new firmware and old

Change-Id: Iab6bc536128689f794eb9694c2503e85835a9ac4

R=sleffler@chromium.org

Review URL: http://codereview.chromium.org/6802011