autotest: add nosave connect parameter for WiFi tests
Add a mechanism for testing the SaveCredentials service property
that controls whether or not a service's security credentials are
retained.  The connect request can now set 'nosave':True to disable
saving of parameters.  This is passed to site_wlan_connect.py who
handles setting the property on the service object.

This will be used in forthcoming tests for SaveCredentials.

BUG=chromium-os:13891
TEST=manual:run new tests and verify the property is set on the service and that no credentials are saved in the profile; also ran MatFunc for regression

Review URL: http://codereview.chromium.org/6823090

Change-Id: I00c58db53efbe34f208e07fc5c5da3140a0763e4