Introduce thermal test for ARM platforms
This CL introduces an autotest to verify the ARM platforms'
thermal management function.

http://codereview.chromium.org/6854025 will add this test to
the autotest set and is required to be able to run the test
procedure described below.

To simulate CPU activity, this test starts multiple
execution threads, all writing 100K bytes of data into a
temp file (one file per thread) in a tight loop.

The new test verifies that the CPU clock speeds up when the
scaling_governor mode is set to 'performance', then it
changes the throttling level to a temperature a couple of
degrees C above the current temperature and the starts more
activity threads and allows for 30 seconds for throttling to
start or for temperature increase by 3C.

Once throttling starts, the test verifies that the CPU clock
is slowed down. Then the test restores the throttling
threshold and observes that throttling is stopped.

In the end of the test (or on any exception) the cleanup()
method is called as an 'after iteration' hook. It stops all running CPU heater threads, deletes all present temp
files created by the threads, and restores where necessary
the scaling governor modes.

Change-Id: If16fa56629e6d35747675ca82861e2cc404a02c5
BUG=chromium-os:12806
TEST=manual

while in chroot, run:

 vbendeb@eskimo ~/trunk/src/scripts $ ./run_remote_tests.sh \
--board=tegra2_kaen --remote=172.22.75.120 power_Thermal

once the test completes successfully, run

 vbendeb@eskimo ~/trunk/src/scripts $ cat \
/tmp/run_remote_tests.latest/power_Thermal/power_Thermal/debug/power_Thermal.INFO:
 04/15 12:00:43 INFO |      test:0288| Test started. Number of iterations: 1
04/15 12:00:43 INFO |      test:0291| Executing iteration 1 of 1
04/15 12:00:43 INFO |      test:0191| Dropping caches between iterations
04/15 12:00:48 INFO |power_Ther:0326| starting temperature is 33.0
04/15 12:00:48 INFO |power_Ther:0327| starting frequency is 1200000
04/15 12:00:48 INFO |power_Ther:0211| changing core 0 governor from ondemand to performance
04/15 12:00:48 INFO |power_Ther:0211| changing core 1 governor from ondemand to performance
04/15 12:00:48 INFO |power_Ther:0106| adding thread number 0
04/15 12:00:48 INFO |power_Ther:0106| adding thread number 1
04/15 12:00:48 INFO |power_Ther:0106| adding thread number 2
04/15 12:00:48 INFO |power_Ther:0106| adding thread number 3
04/15 12:00:48 INFO |power_Ther:0106| adding thread number 4
04/15 12:00:48 INFO |power_Ther:0106| adding thread number 5
04/15 12:00:48 INFO |power_Ther:0106| adding thread number 6
04/15 12:00:48 INFO |power_Ther:0106| adding thread number 7
04/15 12:00:48 INFO |power_Ther:0106| adding thread number 8
04/15 12:00:48 INFO |power_Ther:0106| adding thread number 9
04/15 12:00:58 INFO |power_Ther:0263| full throttle after 10 seconds
04/15 12:00:59 INFO |power_Ther:0340| ending temperature is 35.0
04/15 12:00:59 INFO |power_Ther:0341| ending frequency is 1200000
04/15 12:00:59 INFO |power_Ther:0347| stopping 10 thread(s)
04/15 12:00:59 INFO |power_Ther:0366| restoring core 0 governor to ondemand
04/15 12:00:59 INFO |power_Ther:0366| restoring core 1 governor to ondemand
04/15 12:00:59 INFO |      test:0294| Test finished after 1 iterations.
04/15 12:00:59 INFO |power_Ther:0347| stopping 0 thread(s)

Review URL: http://codereview.chromium.org/6831019