Fixing repeated autotest install on host devices.
Explicitly installing autotest on the remote host before running tests to prevent repeated installations.  The autotest object held by machine_worker does its work on a forked process so the flag saying autotest has been installed on a host is lot on each iteration.  By installing it first we avoid this problem, and if there is ever an issue with the installation on a host autotest verifies the install before running tests and it will still be fixed.

Change-Id: I708f9ed9049047353a40ebd5a680a0a7262f6430

BUG=None.
TEST=BVT runs.

Review URL: http://codereview.chromium.org/6881011