hardware_Backlight: stop powerd to adjust backlight
This test was failing because powerd's auto backlight adjustment was
interfering with the test.  This patch makes the test check for powerd.
If powerd is running, it is turned offm, and later turned back on after
the backlight adjustment test is complete.

BUG=chrome-os-partner:3002
TEST=Make sure you can adjust backlight with the brightness keys.  Start
the test and adjust the backlight continuously while the test runs.
Soon the backlight adjustment controls will not work (powerd stopped).
The backlight should change under the control of the autotest.  When the
test finishes, control should be restored.

For comparison, repeat the same procedure without this patch.  The
backlight test should fail.  Also, try running with this patch but with
powerd stopped.  The test should detect that powerd is not running and
thus not start powerd.  To verify that powerd has not been started at
the end of the test, make sure the brightness cannot be adjusted by
using the brightness keys.

Change-Id: I2c466953803c2701425283d195983e543bd3df9a
Signed-off-by: Simon Que <sque@chromium.org>

Review URL: http://codereview.chromium.org/6871035