This change looks scary, but it shouldn't be once you ignore all the image file changes. Making the browser full screen also helps.
This change does not
a) change any of the existing tests
b) in particular it does not change any timing.

This change should do the following
1) write all images as .png
2) encode pixel MD5 in the image filename
These two things combined make it possible to easily examine images with standard viewers and do a check to classify an image by simply looking it up in a directory of known good/bad images. Also lists of good/bad images can be generated simply by doing an ls on a directory. This means checksums.txt can go away.
3) give more verbose output in case of warning/failures
4) make glbench buildable outside of chroot again controlled via #define
5) copy images into result tree in autotest so they can be easily be looked at when something breaks
6) create a human friendly summary.txt in the result tree for easy failure analysis

The beef of the changes is in testbase.cc and graphics_GLBench.py

The change was tested on mario and tegra.

BUG=chromium-os:14282
TEST=

Review URL: http://codereview.chromium.org/6880025

Change-Id: I480fb7c9c058594c5ccd32b9251a42778296db7f