Remove broken close() call in cros_ui, fixing hardware_ExternalDrives.
In change 93e65d0badb3e0bf25b5c499b683ba6440208d2b, the close() call was
removed from the ChromeSession object.  However, there was still a call
to the close() call in the Dialog.get_entries() function.  That meant
any tests that used Dialog were broken.  For details of the change
that removed the close() call, see:
  http://codereview.chromium.org/6624002

From looking at the original change list (see the comments in change
set #1), it appears that the close() call is simply not needed anymore.
Thus, the fix is simply to remove it.

As part of this change, I have also moved the try/finally for stopping
the http_server to be in a slightly more correct place.  This was done
entirely by code inspection--no failures were actually seen on the
two add_url_handler() calls.

BUG=chrome-os-partner:3441
TEST=semi-automated

Found all applicable tests with:
  cd src/third_party/autotest/files
  git grep cros_ui.Dialog .

Ran tests with:
  ./run_remote_tests.sh --board=x86-mario --remote=${REMOTE_IP} hardware_ExternalDrives
  ./run_remote_tests.sh --board=x86-mario --remote=${REMOTE_IP} graphics_TearTest
  ./run_remote_tests.sh --board=x86-mario --remote=${REMOTE_IP} hardware_USB20
  ./run_remote_tests.sh --board=x86-mario --remote=${REMOTE_IP} hardware_UsbPlugIn

This test fails, though I don't believe it is anything new:
  ./run_remote_tests.sh --board=x86-mario --remote=${REMOTE_IP} hardware_BluetoothSemiAuto

NOTE that the above tests are semi-automated and require user interaction.
Some of them are also VERY slow!

Review URL: http://codereview.chromium.org/6902044

Change-Id: Ifda73b5610fdce4c126fe8e775e8eba44197054b