[autotest] Cleanup log spew from verity autotests
Lots of the shell commands used to set up and check disk images in the DMVerity
autotests spew to stderr even when they succeed.  Add a wrapper that calls
utils.run() in a way that causes this logging to get sent to the DEBUG log.

BUG=chromium-os:4958
TEST=platform_DMVerity*

Change-Id: I718852922438a8a7bcfc71e8ae0c28c973177f0e

R=wad@chromium.org

Review URL: http://codereview.chromium.org/6893064