Surface board host attribute labels for per-board control.
Occasionally we need to disable tests for certain boards. To allow this,
I've surfaced the platform labels for each machine through the standard
site_host_attributes interface.

There's a small catch though. In order to prevent accidental board leaks
we need to make sure names are hashed. As such, the netbook labels
stored in the dict are "netbook_" plus the first eight characters of the
SHA256 hash of the remainder of the netbook label.

Since most people shouldn't be using this feature, a lack of usability
is okay.

For the first test, we need to disable some of ther Kenel crash tests
for a certain board type.

BUG=chromium-os:14497
TEST=Running in production.

Change-Id: Ie006d47fbd6ad898af4858c683065f89e15f5df5