Run hostapd in debugging mode and save log
Add verbose logging to hostapd and copy the log file
after exit and add it to the run_remote_tests file
hierarchy.  This will allow us to diagnose authentication
failures.

In order to capture the whole log, we need to actually
kill hostapd first.  This wasn't happening before in the
error condition.  Change this default behavior, but allow
an override to this behavior on the command line using a
new facility.  While we are here, use this same facility
to also optionally force "skip_test" suites to run.

BUG=n0ne
TEST=Run a few MatFunc tests

Change-Id: I96ef4b30f8de328c23362120025016e751f6072e
Reviewed-on: http://gerrit.chromium.org/gerrit/292
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>