Start WPA_TKIP with negative test
When running autmomated tests in the SecMat test suite,
the first WPA test is WPA_TKIP.  Since the AP is rebooted
at the start of each automated run, and due to the limited
sources of randomness (lack of network traffic?) it appears
to be the case that there isn't enough in /dev/random for
hostapd to use much of the time in the automated testbed.
This means that hostapd ends up saying as follows:

  random: Cannot read from /dev/random: Resource temporarily unavailable
  random: Got 0/20 bytes from /dev/random
  random: Only 0/20 bytes of strong random data available from /dev/random
  random: Not enough entropy pool available for secure operations
  WPA: Not enough entropy in random pool to proceed - reject first 4-way
  handshakeGet randomness: len=32 entropy=11
  [...]
  WPA: Reject 4-way handshake to collect more entropy for random
      number generation

and rejecting the first WPA connection request out-of-hand.  As
a workaround to this behavior, start the WPA_TKIP test with a
negative (bad passphrase) attempt, so that the proceeding attempts
will succeed.

BUG=chromium-os:14930
TEST=Need an automated run to fully test

Change-Id: If48c9d4021cb890646d97966bfc6f85ce13f7e9a
Reviewed-on: http://gerrit.chromium.org/gerrit/328
Reviewed-by: Sam Leffler <sleffler@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>